+++ src/edb_explorer/gui/widgets/query_tab.py
@@ -81,6 +81,8 @@ class ResultsGrid(QWidget):
         layout.addLayout(bar)
         self.view = QTableView()
         self.view.setModel(self.proxy)
+        self.view.setVerticalScrollMode(QAbstractItemView.ScrollMode.ScrollPerPixel)
+        self.view.setHorizontalScrollMode(QAbstractItemView.ScrollMode.ScrollPerPixel)
         self.view.setAlternatingRowColors(True)
         self.view.setSelectionBehavior(QAbstractItemView.SelectionBehavior.SelectRows)
         self.view.setSelectionMode(QAbstractItemView.SelectionMode.ExtendedSelection)